requirements.txt
streamlit
numpy
joblib
pandas
plotly
Pillow
streamlit-drawable-canvas
opencv-python-headless
scikit-learn